feat: add isutility flag table

+++ packages/contracts/src/libraries/LibBuilding.sol
@@ -21,7 +21,7 @@ library LibBuilding {
     uint32 level = 1;
 
     Spawned.set(buildingEntity, true);
-    BuildingType.set(buildingEntity, P_EnumToPrototype.get(BuildingKey, uint32(buildingType)));
+    BuildingType.set(buildingEntity, P_EnumToPrototype.get(BuildingKey, uint8(buildingType)));
     Level.set(buildingEntity, level);
     Position.set(buildingEntity, coord);
     LastClaimedAt.set(buildingEntity, block.number);
@@ -36,7 +36,7 @@ library LibBuilding {
     EBuilding buildingType,
     PositionData memory position
   ) public {
-    bytes32 buildingPrototype = P_EnumToPrototype.get(BuildingKey, uint32(buildingType));
+    bytes32 buildingPrototype = P_EnumToPrototype.get(BuildingKey, uint8(buildingType));
     int32[] memory blueprint = P_Blueprint.get(buildingPrototype);
     Bounds memory bounds = getPlayerBounds(playerEntity);
 
@@ -107,7 +107,7 @@ library LibBuilding {
     EBuilding building,
     uint32 level
   ) internal view returns (bool) {
-    return hasRequiredBaseLevel(playerEntity, P_EnumToPrototype.get(BuildingKey, uint32(building)), level);
+    return hasRequiredBaseLevel(playerEntity, P_EnumToPrototype.get(BuildingKey, uint8(building)), level);
   }
 
   function canBuildOnTile(bytes32 prototype, PositionData memory coord) internal view returns (bool) {
@@ -116,7 +116,7 @@ library LibBuilding {
   }
 
   function canBuildOnTile(EBuilding building, PositionData memory coord) internal view returns (bool) {
-    EResource resource = P_RequiredTile.get(P_EnumToPrototype.get(BuildingKey, uint32(building)));
+    EResource resource = P_RequiredTile.get(P_EnumToPrototype.get(BuildingKey, uint8(building)));
     return resource == EResource.NULL || resource == P_Terrain.get(coord.x, coord.y);
   }
 }
